# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OctaveOneCloud API
  version: 0.2.0
  description: |
    Customer and operator API. Browser clients call it through the web app at `/api/v1/...` (same origin).
    Authentication uses an httpOnly session cookie; state-changing requests must carry a trusted `Origin`.
    Monetary amounts are integers in minor units (paise). Operator endpoints require an operator role and a
    session that has passed TOTP MFA.
servers:
  - url: /api/v1
    description: Through the web app proxy
components:
  securitySchemes:
    session: { type: apiKey, in: cookie, name: ooc_session }
    appSignature: { type: apiKey, in: header, name: x-ooc-signature, description: 'hex(HMAC-SHA256(APP_ADAPTER_<NAME>_SECRET, "<x-ooc-timestamp>.<raw body>")) with headers x-ooc-app and x-ooc-timestamp (unix seconds, within 5 minutes). See docs/app-adapter-contract.md.' }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
        message: { type: string }
        issues: { type: array, items: { type: object, properties: { path: { type: string }, message: { type: string } } } }
        blockers: { type: array, items: { type: string } }
...
  responses:
    Error: { description: Error, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
security: [{ session: [] }]
paths:
...
  /orgs/{orgId}/connectors: { get: { summary: "Configured providers and this organisation's connections (no tokens)", parameters: [{ name: orgId, in: path, required: true, schema: { type: string, format: uuid } }], responses: { '200': { description: OK } } } }
  /orgs/{orgId}/connectors/{provider}/start: { post: { summary: 'Begin OAuth (PKCE); returns { authorizeUrl }', parameters: [{ name: orgId, in: path, required: true, schema: { type: string, format: uuid } }, { name: provider, in: path, required: true, schema: { type: string } }], responses: { '200': { description: OK }, '403': { $ref: '#/components/responses/Error' } } } }
  /orgs/{orgId}/connectors/grants/{grantId}/revoke: { post: { summary: Disconnect, parameters: [{ name: orgId, in: path, required: true, schema: { type: string, format: uuid } }, { name: grantId, in: path, required: true, schema: { type: string, format: uuid } }], responses: { '200': { description: Revoked } } } }
...
  /app-api/entitlements: { post: { security: [{ appSignature: [] }], summary: 'Effective entitlements { orgId }', responses: { '200': { description: OK }, '401': { $ref: '#/components/responses/Error' }, '403': { $ref: '#/components/responses/Error' } } } }
  /app-api/usage/reserve: { post: { security: [{ appSignature: [] }], summary: 'Reserve quota { orgId, resource, quantity, idempotencyKey, ttlSeconds? }', responses: { '200': { description: Reservation }, '409': { description: quota_exceeded } } } }
  /app-api/usage/settle: { post: { security: [{ appSignature: [] }], summary: 'Settle { reservationId, actualQuantity, sourceEventId }', responses: { '200': { description: OK }, '409': { $ref: '#/components/responses/Error' } } } }
  /app-api/usage/release: { post: { security: [{ appSignature: [] }], summary: 'Release { reservationId }', responses: { '200': { description: OK } } } }
  /app-api/approvals/request: { post: { security: [{ appSignature: [] }], summary: 'Request human approval { orgId, actionType, payload, summary, agentRunId?, ttlSeconds? }', responses: { '201': { description: Pending approval } } } }
  /app-api/approvals/status: { post: { security: [{ appSignature: [] }], summary: 'Approval status { orgId, approvalId }', responses: { '200': { description: OK } } } }
  /app-api/approvals/consume: { post: { security: [{ appSignature: [] }], summary: 'Consume once for the identical action { orgId, approvalId, actionType, payload }', responses: { '200': { description: Executed }, '409': { $ref: '#/components/responses/Error' } } } }
  /app-api/connectors/token: { post: { security: [{ appSignature: [] }], summary: 'Customer OAuth access token { orgId, provider } (allowed apps only)', responses: { '200': { description: Token }, '403': { $ref: '#/components/responses/Error' }, '404': { $ref: '#/components/responses/Error' } } } }